+++ src/components/Admin/dasboard/dashboard.js
@@ -13,11 +13,11 @@ import Tooltip from '@mui/material/Tooltip';
 import MenuItem from '@mui/material/MenuItem';
 import {Routes, Route ,Outlet} from "react-router-dom"
 import SideBar from './sidebar';
-import SignIn from '../authentification/signIn';
-import ModalUser from './modal_add_User';
-import ModalHotel from './modal_Add_Hotel';
+import SignIn from '../../authentification/signIn';
+import ModalUser from '../users/modal_add_User';
+import ModalHotel from '../hotels/modal_Add_Hotel';
 // import AllHotels from './AllHotels'
-import AllHotels from './AllHotels'
+import AllHotels from '../hotels/AllHotels'
 import Grid from "@mui/material/Grid"
 
 const settings = ['Profile', 'Account', 'Dashboard', 'Logout'];
