--- src/components/profil.js
@@ -1,18 +1,0 @@
-import React from "react";
-
-
-class Profil extends React.Component {
-
-    constructor(props){
-        super(props)
-    }
-
-    render(){
-        return(
-            <h2>Welcome to profil</h2>
-        )
-    }
-}
-
-
-export default Profil;
